requirements.txt
openai
requests
python-dotenv
schedule
yfinance
